// demo-crud-frontend/src/app/model/User.ts
export interface User {
  id: number;
  about: string;
  address: string;
  age: number;
  balance: string;
  company: string;
  email: string;
  eyeColor: string;
  gender: string;
  isActive: boolean;
  latitude: number;
  longitude: number;
  name: string;
  phone: string;
  picture: string;
  registered: string;
}